Refactor: Update API mutation hooks for Cliente and Usuario; enhance form handling in related components
